Expose Reporter and Input to custom modifiers
This enables custom modifiers to report positioned error messages
