+++ vork/src/main/scala/vork/internal/cli/Settings.scala
@@ -16,23 +16,12 @@ import metaconfig.Configured
 import metaconfig.annotation._
 import metaconfig.generic
 import metaconfig.generic.Surface
+import scala.meta.inputs.Input
 import scala.meta.io.AbsolutePath
 import scala.meta.io.RelativePath
+import vork.CustomModifier
 import vork.Reporter
 import vork.internal.markdown.MarkdownCompiler
-
-trait CustomModifier {
-  val name: String
-  def process(info: String, code: String): String
-}
-
-object CustomModifier {
-  implicit val surface: Surface[Settings] = new Surface(Nil)
-  implicit val decoder: ConfDecoder[CustomModifier] =
-    ConfDecoder.instanceF[CustomModifier](_ => ConfError.message("unsupported").notOk)
-  implicit val encoder: ConfEncoder[CustomModifier] =
-    ConfEncoder.StringEncoder.contramap(mod => s"<${mod.name}>")
-}
 
 case class Settings(
     @Description("The input directory to generate the vork site.")
